+++ libs/gesha-core/src/testing/mod.rs
@@ -3,8 +3,8 @@ pub mod v3_0;
 mod test_case;
 pub use test_case::TestCase;
 
-mod test_case_map;
-pub use test_case_map::{run_parallel, Joiner};
+mod run_parallel;
+pub use run_parallel::{run_parallel, Joiner};
 
 mod test_suite;
 pub use test_suite::TestSuite;
